requirements.txt
# AI Content Pipeline — Python dependencies
# Install with: pip install -r requirements.txt

# Core
python-dotenv>=1.0.0
anthropic>=0.25.0

# Twitter/X (unofficial, no API key needed)
twikit>=2.0.0

# Web scraping & RSS
feedparser>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Video transcription
yt-dlp>=2024.1.0
groq>=0.5.0

# Google Sheets
google-api-python-client>=2.100.0
google-auth>=2.20.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0
